Basic movement
Added images of the ship and basic background images for a parallax
movement.
Messed with the JS some to make it simulate movement through a city,
along with starting the background repeating. This needs work though!
Right now it repeats once to the right but no more. Eventually this
should 'scroll'.

Once the background is sorted, I will move on to projectiles, opponents
and more!