Started work on the real game
The real game is 'sneaky cat', taking the FP-view of a cat, with it's
paws in front. Your aim is to grab things from the table (food, yarn,
etc), while a person in front of you isn't looking. If they are looking,
you lose a life/fail.

Left click - Left paw
Right click - Right paw

There's a lot I'd like to do with this game (Different types of items to
collect, different cats to pick from, multiple levels, dubstep mode,
etc.). It just depends on how much time I can afford for it with so
little time left and so many other assignments.