Slight change in movement
I'm going to make the movement closer to that from Catlateral Damage
(paws unseen, come in from the sides on clicks.. Might make them visible
though, and grow/shrink as they move).
While the rotation was good, it was looking a bit weird and wouldn't
work too well with the way I was trying to make the game work. Instead
they will come in from the sides and pull down straight. It'll make more
sense when 'pushing' items away if a player accidentally clicks the same
paw twice.