New images, more progress on grabbing
More progress on the grabbing animations/logic, with new images to
match.

Still having some issues with rotation, but it's not so bad that it'll
bother me for a while.
I'm working on the logic of 'If left paw is active and right paw is
clicked, put left paw back to normal. If right paw is clicked again,
return right paw to normal'.

It's a little more involved than I thought (that or I'm not giving it
enough focus and should go to sleep <-- Probably this).