The sushi lives!!
I have my first animated sprite on the screen in the form of a cute
little sushi I drew!
Right now it just sits in the middle and blinks. No interaction. The
next step is to get it moving when the user attempts to 'eat' or 'push'
it.

I also did some more work on the logic and now it seems to work just
fine. :) The game is more about 'rhythm' when grabbing items than
mashing of the buttons now. Which is good!