More animations!
The 'item' now moves down from the top to the button on the clicks of
the paw grabbing, and disappears once it reaches the paws.
Not the ideal animation but is alright for my testing purposes.

I want to work on getting multiple items working (probably through the
use of OO style programming), animations for 'eating' and 'knocking' an
item with the paws and the scoring system in play next!

Oh, there's also now an on/off button for the audio. Not pretty, but
alright for my testing. :)