More food, more flinging, more scoring
Score is now kept.
There's now 5 items (Chicken, Sushi, Pancakes, Beef and Ramen).
When an item is 'knocked away', it spins away in a semi-random
direction, based on the paw it was hit with, at varying speeds.
A few more changes.. Also trying to clean things up a bit, although I
expect I'll be doing that right up until the end and then some.