Watcher arrives
Working on the 'loss' condition of the watcher, and how a user is caught
by them as they turn to look toward and away from the player as they are
eating all of the foods!
Also changed the address of the game, just to keep it hidden until
release. :P

Continuing work on the win/loss condition.