Some code refactoring
Mainly moving the code around, making a few new files to make bite sized
chunks of code and sorting things out for allowing the game to start,
end and be paused (with screens for each).

And that's the next step!

After that I'll work on the 'bad' items in the game.
Then perhaps high scores and saving them to local storage?