Closer to the end..
The game now has some recognition of the 'Loss', and restarts to the
main menu after a grey screen when all lives are lost.

There are now also menu options for the highscores, help and credits
sections, although no content is in them yet you can click and go to
them and then click back to the menu.

Still need to do the 'bad' items, multipliers and potentially levels..