Keeping Score
The game now keeps the current highscore and also saves the overall high
score for the player (on that particular browser), in HTML5 local
storage. Horay.

Still need to fill out the Help and Credits menu, and work on the 'bad'
items and potentially levels.

I think the next step is the bad items. I can't do the 'Help' menu
before then.