Bad items, preloading images, stuff!
Now includes the 'bad' items of a bomb and a 1 tonne weight :D.

Also has pre-loading of images and has been moved to 'index'.
Thinking about scrapping the need for levels. It sucks, but time is
running short.

Left to do:
Help page
About page
Alternate Keyboard controls
I think that's it for the main bit??