Game and website finished
For now, I'm happy to call this finished.

There are a lot of things I would like to do with it, had I more time,
but what it is isn't bad. It fits the assignment spec and then some, and
I'm happy with the result.

I'll talk about the things I wanted to add but didn't/couldn't in the
report that I'll be writing tomorrow. :) I -think- everything else for
the assignment marks are done though. Horay. I won't submit it until
I've done the 'testing' section of my report though either. Just in case
:D