Finished both parts
I now consider this assignment finished! (Yet to upload to the
submission, but I'll do that after uploading this).

I've made the game run at generally the same speed on all browsers now.
The sound effects on IE can now be played more than once per effect.

The written aspect is done. It's not the best I've done, and I felt
sorely limited by the 1000 word count, but I'm okay with where it is,
and think it sums up each heading as much as needed.

Another one checked off! I enjoyed this assignment, and am thinking of
maybe porting my Java game over to JavaScript for an in-browser game of
that sort. Maybe anyway. I wouldn't mind working with canvas again to
make more games. :D